pyproject.toml
[tool.poetry]
name = "gpu-use"
version = "0.1.0"
description = ""
authors = ["Erik Wijmans <etw@gatech.edu>"]

[tool.poetry.scripts]
gpu-used = 'gpu_use:run_daemon'
gpu-use = 'gpu_use:gpu_use_cli'

[tool.poetry.dependencies]
python = ">=3.6.1"
sqlalchemy = "^1.3.17"
click = "^7.1.2"
colorama = "^0.4.3"
mysqlclient = "^1.4.6"
python-daemon = "^2.2.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pre-commit = "^2.4.0"
python-language-server = "^0.32.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
